making scdoc optional

+++ PKGBUILD
@@ -6,7 +6,8 @@ pkgdesc="A tiling layout generator for river"
 arch=('x86_64')
 url="https://gitlab.com/snakedye/kile"
 license=('MIT')
-makedepends=('git' 'rust' 'scdoc')
+makedepends=('git' 'rust')
+optdepends=('scdoc')
 provides=("kile-wl")
 conflicts=("kile-wl")
 source=("$pkgname::git+https://gitlab.com/snakedye/kile.git")
@@ -20,8 +21,6 @@ pkgver() {
 build() {
   cd "$pkgname"
   cargo build --release
-  cd doc
-  scdoc < kile.1.scd > kile.1.gz
 }
 
 package() {
@@ -29,6 +28,8 @@ package() {
   mkdir -p "$pkgdir/usr/share/man/man1"
 	cd "$pkgname"
   install -Dm755 target/release/kile "$pkgdir/usr/bin/"
-  install -Dm644 doc/*.1.gz "$pkgdir/usr/share/man/man1/"
+  if [ -e doc/*.1.gz ]; then
+    install -Dm644 doc/*.1.gz "$pkgdir/usr/share/man/man1/"
+  fi
 }
